feat / Complete templating and styling for chat
Updated sidebar table controller to view controller and revamped design of
sidebar to be a chat menu and any clicks to enter a chat lobby with targeted
user. All pages have been styled properly but lack all the core functionality.
Placed the tableview and text entry point inside of a scroll view in order to
solve keyboard showing up and covering the input field.

Commits:
- Overhauled right reveal bar to be chat menu
- Complete styling for right reveal chat bar
- Created redirect when a person is clicked on chat
- Complete templating by placing tableview and text field inside scroll view